requirements.txt
fastapi
uvicorn
pdfplumber
pydantic
python-multipart
python-dotenv
google-generativeai